Ensure all phase and survey data are committed before advancing or uploading; move OSF upload to completion page for data integrity
